refactor resolvers into separate modules
