+++ src/index.js
@@ -1,179 +1,21 @@
 import { GraphQLServer } from "graphql-yoga";
-import uuidv4 from "uuid/v4";
 
 import db from "./db";
-
-const resolvers = {
-  Query: {
-    me() {
-      return {
-        id: "3662",
-        name: "Emma",
-        email: "emma@example.com",
-        age: 29
-      };
-    },
-    users(parent, args, { db }, info) {
-      if (!args.query) {
-        return db.users;
-      }
-      return db.users.filter(user =>
-        user.name.toLowerCase().includes(args.query.toLowerCase())
-      );
-    },
-    post() {
-      return {
-        id: "1",
-        title: "First Post",
-        body: "This is the first post",
-        published: true,
-        author: "3662"
-      };
-    },
-    posts(parent, args, { db }, info) {
-      if (!args.query) {
-        return db.posts;
-      }
-      return db.posts.filter(post => {
-        const query = args.query.toLowerCase();
-        const hasTitleMatch = post.title.toLowerCase().includes(query);
-        const hasBodyMatch = post.body.toLowerCase().includes(query);
-        return hasTitleMatch || hasBodyMatch;
-      });
-    },
-    comments() {
-      return db.comments;
-    }
-  },
-  Mutation: {
-    createUser(parent, args, { db }, info) {
-      const emailTaken = db.users.some(user => user.email === args.data.email);
-
-      if (emailTaken) {
-        throw new Error("Email already in use");
-      }
-
-      const user = {
-        id: uuidv4(),
-        ...args.data
-      };
-
-      db.users.push(user);
-      return user;
-    },
-    deleteUser(parent, args, { db }, info) {
-      const userIndex = db.users.findIndex(user => user.id === args.id);
-
-      if (userIndex === -1) {
-        throw new Error("User not found");
-      }
-
-      const deletedUsers = db.users.splice(userIndex, 1);
-
-      db.posts = db.posts.filter(post => {
-        const match = post.author === args.id;
-
-        if (match) {
-          db.comments = db.comments.filter(comment => comment.post !== post.id);
-        }
-
-        return !match;
-      });
-      db.comments = db.comments.filter(comment => comment.author !== args.id);
-
-      return deletedUsers[0];
-    },
-    createPost(parent, args, { db }, info) {
-      const authorExists = db.users.some(user => user.id === args.data.author);
-
-      if (!authorExists) {
-        throw new Error("Author not found");
-      }
-
-      const post = {
-        id: uuidv4(),
-        ...args.data
-      };
-
-      db.posts.push(post);
-      return post;
-    },
-    deletePost(parent, args, { db }, info) {
-      const postIndex = db.posts.findIndex(post => post.id === args.id);
-
-      if (postIndex === -1) {
-        throw new Error("Post not found");
-      }
-
-      const deletedPosts = db.posts.splice(postIndex, 1);
-
-      db.comments = db.comments.filter(comment => comment.post !== args.id);
-
-      return deletedPosts[0];
-    },
-    createComment(parent, args, { db }, info) {
-      const authorFound = db.users.some(user => user.id === args.data.author);
-      const postFound = db.posts.some(
-        post => post.id === args.data.post && post.published
-      );
-
-      if (!authorFound) {
-        throw new Error("Comment author not found");
-      } else if (!postFound) {
-        throw new Error("Post not found");
-      }
-
-      const comment = {
-        id: uuidv4(),
-        ...args.data
-      };
-
-      db.comments.push(comment);
-
-      return comment;
-    },
-    deleteComment(parent, args, { db }, info) {
-      const commentIndex = db.comments.findIndex(
-        comment => comment.id === args.id
-      );
-
-      if (commentIndex === -1) {
-        throw new Error("Comment not found");
-      }
-
-      const deletedComments = db.comments.splice(commentIndex, 1);
-      return deletedComments[0];
-    }
-  },
-  Post: {
-    author(parent, args, { db }, info) {
-      return db.users.find(user => user.id === parent.author);
-    },
-    comments(parent, args, { db }, info) {
-      return db.comments.filter(comment => comment.post === parent.id);
-    }
-  },
-  User: {
-    posts(parent, args, { db }, info) {
-      return db.posts.filter(post => post.author === parent.id);
-    },
-    comments(parent, args, { db }, info) {
-      return db.comments.filter(comment => comment.author === parent.id);
-    }
-  },
-  Comment: {
-    author(parent, args, { db }, info) {
-      return db.users.find(user => user.id === parent.author);
-    },
-    post(parent, args, { db }, info) {
-      return db.posts.find(post => post.id === parent.post);
-    }
-  }
-};
+import Query from "./resolvers/Query";
+import Mutation from "./resolvers/Mutation";
+import User from "./resolvers/User";
+import Post from "./resolvers/Post";
+import Comment from "./resolvers/Comment";
 
 const server = new GraphQLServer({
   typeDefs: "./src/schema.graphql",
-  resolvers,
+  resolvers: {
+    Query,
+    Mutation,
+    User,
+    Post,
+    Comment
+  },
   context: {
     db
   }
