+++ src/resolvers/Comment.js
@@ -1,0 +1,10 @@
+const Comment = {
+  author(parent, args, { db }, info) {
+    return db.users.find(user => user.id === parent.author);
+  },
+  post(parent, args, { db }, info) {
+    return db.posts.find(post => post.id === parent.post);
+  }
+};
+
+export default Comment;
